[Feature/#17] : 날씨, 인구 혼잡도 api 연동 및 cleartext 이슈 해결

+++ data/src/main/java/com/gdg/data/datasource/CrowdZeroDataSource.kt
@@ -5,6 +5,6 @@ import com.gdg.data.dto.response.CongestionResponseDto
 import com.gdg.data.dto.response.WeatherResponseDto
 
 interface CrowdZeroDataSource {
-    suspend fun getWeather(areaId: Long): BaseResponse<WeatherResponseDto>
-    suspend fun getCongestion(areaId: Long): BaseResponse<CongestionResponseDto>
+    suspend fun getWeather(areaId: Int): BaseResponse<WeatherResponseDto>
+    suspend fun getCongestion(areaId: Int): BaseResponse<CongestionResponseDto>
 }
